# docs/operations/api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Thesis Engine API",
    "description": "Backend service layer exposing PTE artefacts (canonical state, valuation, forecast, ficha, peers, cross-check logs) and the yaml authoring + pipeline-trigger workflow.",
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/tickers/{ticker}/yamls": {
      "get": {
        "tags": [
          "yamls"
        ],
        "summary": "List Yamls",
        "description": "List the analyst-editable yamls present for ``ticker``.",
        "operationId": "list_yamls_api_tickers__ticker__yamls_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Yamls Api Tickers  Ticker  Yamls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/yamls/{yaml_name}": {
      "get": {
        "tags": [
          "yamls"
        ],
        "summary": "Download Yaml",
        "description": "Raw yaml text for the requested file.",
        "operationId": "download_yaml_api_tickers__ticker__yamls__yaml_name__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "yaml_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Yaml Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "yamls"
        ],
        "summary": "Upload Yaml",
        "description": "Upload yaml content. 422 with structured errors on validation\nfailure; 200 with backup path + new filename on success.",
        "operationId": "upload_yaml_api_tickers__ticker__yamls__yaml_name__post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "yaml_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Yaml Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "title": "Content"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Yaml Api Tickers  Ticker  Yamls  Yaml Name  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickers/{ticker}/yamls/{yaml_name}/versions": {
      "get": {
        "tags": [
          "yamls"
        ],
        "summary": "List Versions",
        "description": "Backup history for a yaml \u2014 most recent first.",
        "operationId": "list_versions_api_tickers__ticker__yamls__yaml_name__versions_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "yaml_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Yaml Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Versions Api Tickers  Ticker  Yamls  Yaml Name  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}